requirements.txt
opencv-python
numpy
scipy
pyserial
imutils
time
kivy
kivymd
matplotlib
kivy_garden.matplotlib
setuptools
